+++ changelog.md
@@ -1,5 +1,6 @@
 # v5.4.2-dev
 - Remove dependency of `mplutils`
+- Add `Hist1d/2d.copy()`
 
 # v5.4.1
 - Expose `for_pcolormesh_from_txt` and `for_pcolormesh_from_root`